Init  splitting
